tutor phases WIP
